pyproject.toml
[project]
name = "dashbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Oscar Ek"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "decouple (>=0.0.7,<0.0.8)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "mypy (>=1.18.1,<2.0.0)",
    "newspaper3k (>=0.2.8,<0.3.0)",
    "google-api-python-client (>=2.0.0,<3.0.0)",
    "lxml-html-clean (>=0.4.2,<0.5.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "openai (>=1.107.3,<2.0.0)",
    "boto3 (>=1.35.0,<2.0.0)",
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-asyncio (>=0.24.0,<1.0.0)",
    "pytest-mock (>=3.12.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
